# api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Internal HTTP API",
    "version": "1.0.0-alpha",
    "summary": "Public reader endpoints, editorial writes behind a bearer JWT, and the operational probes served by the Go binary.",
...
  },
  "servers": [
    {
      "url": "http://api:8080",
      "description": "The internal service address inside compose and Kubernetes; the Astro server reaches it as API_BASE_URL."
    },
    {
      "url": "http://localhost:8080",
      "description": "A locally run binary (HTTP_ADDR defaults to :8080)."
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/v1/cashback/catalogue": {
      "get": {
        "tags": [
          "cashback-member"
        ],
        "operationId": "listCatalogue",
        "summary": "The retailers a member can earn at, one page at a time",
        "description": "The catalogue as a shelf: every active retailer in the places the reader follows, described in the reader's language, each with the rate bands in force for them right now (US5 scenario 1, FR-010).\n\n**Every rate here is what the MEMBER earns**, never the network's commission - the same bands, resolved by the same reader and rendered in the same shape as `GET /merchants/{slug}`, so the shelf and the shop page cannot disagree between two clicks.\n\n**Language and place are independent axes** (constitution VII). `lang` is required and names the language only: a retailer with no copy in it is shown in their source language and says so in `name_is_fallback`. `place` is required and repeatable; a place scopes to itself and everything under it, so `germany` lists what `munich` lists and more. A `place` that names nothing is refused rather than dropped, because a shelf scoped to fewer places than were asked for looks exactly like a shelf.\n\n**`q` searches the name the reader sees** - the resolved one, fallback included - so a Greek reader searching for a retailer shown to them in German finds it under the name on their screen.\n\n**Pages run in slug order.** `next_cursor` is null on the last page. A retailer imported between two pages appears on a later page if its slug sorts later, and not at all in this walk if it sorts earlier; the next walk sees it.\n\n**`category` is refused, not ignored.** Nothing in the schema records one yet (#414), and a filter that is silently ignored is a page of the wrong shops with no error on it.\n\nAn empty shelf is 200 with an empty `items`, never a 404. The whole path is unmounted when cashback is off, and then answers 404.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "description": "The reader's language, as a BCP-47 primary subtag. Required: a shelf of forty names in a language the reader did not choose is a shelf they cannot use. Copy missing in it falls back to each retailer's source language, labelled.",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z]{2,3}$",
              "examples": [
                "el",
                "de"
              ]
            }
          },
          {
            "name": "place",
            "in": "query",
            "required": true,
            "description": "Slug of a place the reader follows. Repeat the parameter to shelve several places at once (`?place=munich&place=greece`). Each scopes to itself and everything under it. A slug that names no place is refused with 400.",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "minItems": 1,
              "items": {
                "type": "string",
                "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
              }
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Free text matched, case-insensitively, against the retailer name the reader would see on this shelf.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size. Absent means the default; above the maximum is capped to it rather than refused, because a client that asked for too much wanted a page. Zero, negative or not a whole number is refused with 400.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque page cursor: pass back the `next_cursor` of the previous page. Its encoding is an implementation detail and may change; clients must treat it as bytes. One this listing did not issue, or over 256 characters, is refused with 400.",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "deprecated": true,
            "description": "Named by the contract and not yet built (#414). Any value, including an empty one, is refused with 400 rather than ignored.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of the shelf.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CataloguePage"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed `lang`, missing or unknown `place`, a `category`, a `limit` that is not a positive whole number, or a `cursor` this listing did not issue. The detail says which; an unknown place is named.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                },
                "examples": {
                  "noLang": {
                    "summary": "No language",
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "lang is required: the reader's language as a BCP-47 primary subtag, such as el or de"
                    }
                  },
                  "unknownPlace": {
                    "summary": "A place that names nothing",
                    "value": {
                      "type": "about:blank",
                      "title": "Bad Request",
                      "status": 400,
                      "detail": "no such place: \"atlantis\""
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/cashback/merchants/{slug}": {
      "get": {
        "tags": [
          "cashback-member"
        ],
        "operationId": "getMerchant",
        "summary": "One retailer's page, with every rate band published for them",
        "description": "The retailer at this address, described in the reader's language, with EVERY rate band in force right now - each carrying its own conditions and its own exclusions (US5 scenario 3, FR-011).\n\n**Every rate here is what the MEMBER earns**, never the network's commission. A published band records the commission and, separately, the share of it the member receives; this endpoint composes the two before answering, rounding the way the credit itself will round, so the figure shown and the figure later paid can only differ over the commission the network actually reported.\n\n**Bands are the published route's.** A retailer live on several networks has a band per network; the catalogue publishes one route, which is the route a click is issued through. Quoting the others would quote a rate no click can earn.\n\n**An empty `rates` list is an ordinary answer.** A retailer whose bands have lapsed pays nothing today and is still a shop that exists, so it is 200 with no rates - never a 404.\n\n**Language and place are independent axes** (constitution VII). `lang` names the language only. A merchant with no copy in it falls back to the retailer's source language and says so in `name_is_fallback` - an unrecognised or absent `lang` does the same rather than failing, because a page a member has navigated to should render.\n\nThe whole path is unmounted when cashback is off, and then answers 404.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "The retailer's stable, human-readable address.",
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "description": "BCP-47 primary language subtag the reader wants the copy in. Absent or unavailable falls back to the retailer's source language, labelled.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The retailer's page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Merchant"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No retailer is published at this address: never imported, or imported and since paused. The two are one answer deliberately - which retailers we have stopped publishing is not a page request's business."
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A JWT from the auth provider, verified against the configured JWKS. The subject must map to an account. Endpoints under /api/v1/editorial/ additionally require that account to hold the editor role, and the database enforces that rule again on write; endpoints under /api/v1/account/ require only that somebody is signed in, because they act on the caller's own row."
      }
    },
...
    "responses": {
      "Unauthenticated": {
        "description": "No bearer token, or one that is invalid, expired or belongs to no account.",
        "headers": {
          "WWW-Authenticate": {
            "description": "`Bearer realm=\"editorial\"`, with `error=\"invalid_token\"` appended when a token was supplied but rejected.",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "examples": {
              "missingToken": {
                "value": {
                  "type": "about:blank",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "a bearer token is required"
                }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "The request failed for a reason the caller cannot act on. Internals stay in the log: the body carries no database or stack detail, and `detail` may be absent entirely.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "examples": {
              "opaque": {
                "value": {
                  "type": "about:blank",
                  "title": "Internal Server Error",
                  "status": 500
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "title": "Problem",
        "description": "RFC 9457 problem details, as every handler in this API writes them.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "description": "Always `about:blank`, which per RFC 9457 makes `title` the plain rendering of the status code. No problem-type URIs are minted for this API.",
            "const": "about:blank"
          },
          "title": {
            "type": "string",
            "description": "The HTTP status text, e.g. `Bad Request`.",
            "examples": [
              "Bad Request"
            ]
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code, repeated in the body.",
            "minimum": 400,
            "maximum": 599,
            "examples": [
              400
            ]
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation of this occurrence. Omitted when the server has nothing safe to say (notably on 500)."
          }
        }
      },
...
      "Money": {
        "type": "object",
        "description": "An amount of money: whole minor units and the ISO-4217 code they are denominated in. Never a decimal and never a float - the frontend formats these two fields for a locale, and nothing between here and there has to agree on how many decimal places a currency has (C-6).",
        "required": [
          "minor",
          "currency"
        ],
        "additionalProperties": false,
        "properties": {
          "minor": {
            "type": "integer",
            "format": "int64",
            "description": "The amount in minor units, e.g. 4999 for 49.99 EUR."
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "example": "EUR"
          }
        }
      },
...
      "CatalogueItem": {
        "type": "object",
        "description": "One retailer on the shelf: the shop page's shape minus what only the page shows (the country, the terms, the confirmation estimate). A client renders a card from it and follows `slug` to `GET /merchants/{slug}` for the rest.",
        "required": [
          "merchant_id",
          "slug",
          "name",
          "name_language",
          "name_is_fallback",
          "summary",
          "rates"
        ],
        "properties": {
          "merchant_id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "description": "The retailer's address: what `GET /merchants/{slug}` takes."
          },
          "name": {
            "type": "string"
          },
          "name_language": {
            "type": "string",
            "description": "The language the copy on this card is actually in, which is not necessarily the one that was asked for."
          },
          "name_is_fallback": {
            "type": "boolean",
            "description": "Whether the copy came from the retailer's source language because none exists in the reader's. This is how US5 scenario 2 renders \"shown in German\" rather than silently pretending."
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "rates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MerchantRate"
            },
            "description": "Every band in force for this retailer right now, newest first, in the shape the shop page publishes them. Empty when their rates have lapsed, which is an ordinary answer."
          }
        }
      },
      "CataloguePage": {
        "type": "object",
        "description": "One page of the shelf.",
        "required": [
          "items",
          "next_cursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogueItem"
            },
            "description": "The retailers on this page, in slug order. Empty when the places hold no retailer yet."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pass back as `cursor` for the next page. Null on the last page."
          }
        }
      },
      "Merchant": {
        "type": "object",
        "description": "A retailer's page: who they are, how they are described in the reader's language, and every rate band published for them at the moment the page was read.",
        "required": [
          "merchant_id",
          "slug",
          "country",
          "name",
          "name_language",
          "name_is_fallback",
          "summary",
          "terms",
          "typical_confirmation_days",
          "rates"
        ],
        "properties": {
          "merchant_id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "description": "The address this page was reached at."
          },
          "country": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[A-Z]{2}$",
            "description": "Where the retailer trades, or null when they are bound to no one country. Not where the reader is: language and place are independent axes (constitution VII)."
          },
          "name": {
            "type": "string"
          },
          "name_language": {
            "type": "string",
            "description": "The language the copy on this page is actually in, which is not necessarily the one that was asked for."
          },
          "name_is_fallback": {
            "type": "boolean",
            "description": "Whether the copy came from the retailer's source language because none exists in the reader's. This is how US5 scenario 2 renders \"shown in German\" rather than silently pretending."
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "terms": {
            "type": [
              "string",
              "null"
            ]
          },
          "typical_confirmation_days": {
            "type": [
              "integer",
              "null"
            ],
            "description": "How long a purchase here usually takes to confirm. ALWAYS null today: nothing in the schema records it - not on the retailer, not on the route, not on the network - and a member reads this number as \"you will be paid in about six weeks\", so it is left absent rather than filled in with a plausible constant."
          },
          "rates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MerchantRate"
            },
            "description": "Every band in force right now, newest first. Empty when the retailer's rates have lapsed, which is an ordinary answer rather than a 404."
          }
        }
      },
      "MerchantRate": {
        "type": "object",
        "description": "One published rate band, as the member earns it. Exactly one of `bps` and `amount` is present, chosen by `kind`; both are integers, because no rate or money value on this API is ever a decimal string or a float (C-6). The frontend formats for display from these fields.",
        "required": [
          "offer_id",
          "kind",
          "conditions",
          "exclusions",
          "valid_to"
        ],
        "properties": {
          "offer_id": {
            "type": "string",
            "format": "uuid",
            "description": "The band a click on this rate would be issued against - the value POST /clickouts takes, so a page and the click it leads to name the same rate."
          },
          "kind": {
            "type": "string",
            "enum": [
              "percent",
              "fixed"
            ],
            "description": "Which of `bps` and `amount` carries this rate."
          },
          "bps": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000,
            "description": "The member's rate in basis points on a percent band (4% = 400). Absent on a fixed band."
          },
          "amount": {
            "$ref": "#/components/schemas/Money",
            "description": "What the member earns per transaction on a fixed band. Absent on a percent band."
          },
          "conditions": {
            "type": [
              "string",
              "null"
            ],
            "description": "What this band requires, verbatim as published (FR-011), or null when it records none. Null rather than an empty string, so \"no conditions\" and \"conditions that failed to load\" do not render the same."
          },
          "exclusions": {
            "type": [
              "string",
              "null"
            ],
            "description": "What this band does not pay on (FR-011), or null when it records none."
          },
          "valid_to": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When this band stops being published, or null for one with no published end."
          }
        }
      },
...
    }
  }
}